requirements.txt
Flask==0.12.2
RPi.GPIO==0.6.3
python-vlc==2.2.6100
rpi-ws281x==1.0.0
smbus==1.1
